+++ epc/schemes/__init__.py
@@ -2,7 +2,8 @@ from .GID import GID
 from .GIAI import GIAI
 from .GRAI import GRAI
 from .SGLN import SGLN
+from .SGTIN import SGTIN
 
 __all__ = (
-    'GID', 'GIAI', 'GRAI', 'SGLN',
+    'GID', 'GIAI', 'GRAI', 'SGLN', 'SGTIN'
 )