+++ epc/utils/upc.py
@@ -4,16 +4,16 @@ GTIN_MATCH = re.compile(r'^\d{8}$|^\d{12,14}$')
 
 
 def calc_check_digit(upc_string):
-    odd_sum = 0
-    even_sum = 0
+    evens = []
+    odds = []
 
     for i, char in enumerate(upc_string):
-        if (i + 1) % 2 == 0:
-            even_sum += int(char)
+        if (i + 2) % 2 == 0:
+            evens.append(int(char))
         else:
-            odd_sum += int(char) * 3
+            odds.append(int(char))
 
-    return 10 - ((odd_sum + even_sum) % 10)
+    return (10 - (((3 * sum(evens)) + sum(odds)) % 10)) % 10
 
 
 def is_valid_upc(upc_string):