+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+### v1.2
+
+- Added SGTIN encoding support.
+- Fixed sgln-195 encoding not being correctly detected when using `decode_epc()`.
+
+
 ### v1.1
 
 - Fixed incorrect SGLN check digit calculation if the sum of digits was a multiple of 10. Thanks @endreszabo!